revisi layout

+++ database/migrations/2024_07_18_131818_create_nilai_alternatifs_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('alternatif_id')->constrained('alternatifs')->onDelete('restrict')->onUpdate('cascade');
             $table->foreignId('kriteria_id')->constrained('kriterias')->onDelete('restrict')->onUpdate('cascade');
-            $table->float('nilai');
+            $table->double('nilai');
             $table->timestamps();
         });
     }
